refactor(driver): 将internal/driver目录移动到internal/protocol/mysql目录下 (#28)
