+++ internal/protocol/mysql/driver/log/connector.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"database/sql/driver"
 
-	driver2 "github.com/meoying/dbproxy/internal/driver"
+	driver2 "github.com/meoying/dbproxy/internal/protocol/mysql/driver"
 )
 
 var _ driver2.Connector = &connectorWrapper{}
