+++ internal/protocol/mysql/driver/log/stmt.go
@@ -5,7 +5,7 @@ import (
 	"database/sql/driver"
 
 	"github.com/ecodeclub/ekit/slice"
-	driver2 "github.com/meoying/dbproxy/internal/driver"
+	driver2 "github.com/meoying/dbproxy/internal/protocol/mysql/driver"
 )
 
 var _ driver2.Stmt = &stmtWrapper{}
