+++ internal/protocol/mysql/driver/log/transaction.go
@@ -3,7 +3,7 @@ package log
 import (
 	"database/sql/driver"
 
-	driver2 "github.com/meoying/dbproxy/internal/driver"
+	driver2 "github.com/meoying/dbproxy/internal/protocol/mysql/driver"
 )
 
 var _ driver2.Tx = &txWrapper{}
